Refactor to share assertions via test-dist.test.ts and delete test-utils.ts
