+++ harness/tests/copy-skills.test.ts
@@ -5,6 +5,7 @@ import path from 'path';
 import { execSync } from 'child_process';
 import { createIsolatedHome, copySkills, cleanupIsolatedHome } from '../lib/agent-shared.ts';
 import { Agents } from '../config.ts';
+import { assertSearchResults } from '../../serving/skills-cli/test-dist.test.ts';
 
 test('copySkills sets up the isolated environment with the skill and its data', async () => {
     let homeDir = '';
@@ -41,19 +42,7 @@ test('copySkills sets up the isolated environment with the skill and its data', 
         // We need to extend PATH to make sure node is available if needed, but it should be
         const cmd = `node ${mjsPath} --search "address form"`;
         const output = execSync(cmd, { encoding: 'utf8' });
-        const results = JSON.parse(output);
-        assert.ok(Array.isArray(results), 'Output should be a JSON array');
-        assert.ok(results.length > 0, 'Search should find some results');
-        
-        // Find if 'autofill-address-form' is in the results
-        const hasAddressForm = results.some((r: any) => r.id === 'autofill-address-form');
-        assert.ok(hasAddressForm, 'Results should contain autofill-address-form');
-        
-        // Verify structure of the first item
-        const topResult = results[0];
-        assert.ok(topResult.id, 'Top result should have an id');
-        assert.ok(topResult.description, 'Top result should have a description');
-        assert.ok(topResult.distance, 'Top result should have a distance');
+        assertSearchResults(output);
         
     } finally {
         if (homeDir) {
