+++ serving/skills-cli/test-dist.test.ts
@@ -3,16 +3,34 @@ import assert from 'node:assert';
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import { execSync } from 'node:child_process';
-import { assertSearchResults } from './test-utils.ts';
+import { fileURLToPath } from 'node:url';
+
+export function assertSearchResults(output: string) {
+    const results = JSON.parse(output);
+    assert.ok(Array.isArray(results), 'Output should be a JSON array');
+    assert.ok(results.length > 0, 'Search should find some results');
+    
+    // Find if 'autofill-address-form' is in the results
+    const hasAddressForm = results.some((r: any) => r.id === 'autofill-address-form');
+    assert.ok(hasAddressForm, 'Results should contain autofill-address-form');
+    
+    // Verify structure of the first item
+    const topResult = results[0];
+    assert.ok(topResult.id, 'Top result should have an id');
+    assert.ok(topResult.description, 'Top result should have a description');
+    assert.ok(topResult.distance, 'Top result should have a distance');
+}
 
 const ROOT_DIR = path.resolve(import.meta.dirname, "../.."); // guidance/
 const DIST_DIR = path.join(ROOT_DIR, "dist/skills-cli");
 
-console.log("Running build-dist to ensure fresh build...");
-execSync('node --experimental-strip-types skills-cli/build-dist.ts', { 
-  cwd: path.resolve(import.meta.dirname, '..'), 
-  stdio: 'inherit' 
-});
+if (process.argv[1] === fileURLToPath(import.meta.url) || process.argv[1]?.endsWith('test-dist.test.ts')) {
+    console.log("Running build-dist to ensure fresh build...");
+    execSync('node --experimental-strip-types skills-cli/build-dist.ts', { 
+      cwd: path.resolve(import.meta.dirname, '..'), 
+      stdio: 'inherit' 
+    });
+}
 
 
 test('Dependency parity across package.json manifests', async () => {
